feat(server) - fetching trending (popular, on the air and top rated) tv shows from TheMovieDB API

+++ packages/server/src/graphql/typeDefs/types/movie.ts
@@ -115,7 +115,7 @@ export default gql`
     poster_path: String
     popularity: Float
     original_language: String
-    vote_count: Float
+    vote_count: Int
     id: Int
   }
